ARDd: Channel Multiplexer, start incomplete unit test
Channel multiplexing is designed to transport many communication channels
onto a single channel.
For example, many InputStreams/OutputStreams can be used on a single Socket.
This way, only one Socket is needed instead of many.

This is needed in order to address the message blocking issues in the
ARDBroker. If one end is blocking because it's waiting for a reply, then it
also means that end cannot receive any other messages that are non-dependent
on the one being waited on. This gets in the way of being able to cancel
commands or communicating concurrently.

git-svn-id: https://itsd.ict.sait.ca:9880/capstone07/trunk@127 0c6396e7-dfa7-ad41-aac3-ea9b0ac9ed18