WEB: Change brokers to only accept a DataSource and implement AutoCloseable
As it turns out, accepting a javax.sql.Connection causes problems.
Will the broker close the connection on the programmer's behalf?
If not, why should it be passed to the broker and not a DataSource
(which creates a Connection)?

It also causes problems with being able to close the broker. AutoCloseable
was implemented in order to make using the brokers simpler in the web
controllers. If the programmer has to make guesses about whether close() in
a broker interface will gravely screw with the state of the program, there's
something wrong.

git-svn-id: https://itsd.ict.sait.ca:9880/capstone07/trunk@148 0c6396e7-dfa7-ad41-aac3-ea9b0ac9ed18