CONSTANTS: genconst.py - A tool for generating constants usable in Java and MySQL
This is intended to avoid maintaining insufferable amounts of Java boilerplate
and to centralize constants used in many domains. For example, the maximum
username length is used in MySQL as well as Java (Controllers, JSPs).
When a single constant is updated, the programmer shouldn't have to hunt
down other references to it and change those too.

git-svn-id: https://itsd.ict.sait.ca:9880/capstone07/trunk@247 0c6396e7-dfa7-ad41-aac3-ea9b0ac9ed18